requirements.txt
asgiref==3.8.1
Django==5.0.8
sqlparse==0.4.3
tzdata==2022.6
gunicorn==20.1.0
psycopg2==2.9.9
dj-database-url==2.2.0
requests==2.32.3
selenium==4.23.1
djangorestframework==3.15.2
django-cors-headers==4.4.0